requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
exceptiongroup==1.2.0
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
jwcrypto==1.5.6
packaging==24.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pytest==8.1.1
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.28
sqlmodel==0.0.16
starlette==0.36.3
tomli==2.0.1
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
